In preparation for settings refactoring, various refactorings:
- Moved non-class files out of src
- Moved Settings into src.
- Cleaned up lint findings
- Minor adjustments to logic in edit_config_dialog
- Key parser as utility class

tests should still pass, and playing a game seems to still work :)
