+++ src/executer_pyautogui.py
@@ -1,8 +1,9 @@
 import pyautogui
+import constants
+import utilities
 from src.executer_base import BaseExecutor
-from src import utilities, constants
 from src.view.view_base import SettingsItem
-from src.classes.settings import Settings
+from src.settings import Settings
 
 TRIGGER_DELAY = "pyautogui_triggerDelay"
 TRIGGER_DELAY_DEFAULT = 100
