+++ src/key_parser_pynput.py
@@ -1,17 +1,19 @@
 from pynput import keyboard
 
-def parse_key(key):
-    if key in key_map:
-        return key_map[key]
-    else:
-        return key
+class PynputKeyparser:
+    key_map = {
+        "shift":keyboard.Key.shift,
+        "ctrl":keyboard.Key.ctrl,
+        "up":keyboard.Key.up,
+        "down":keyboard.Key.down,
+        "left":keyboard.Key.left,
+        "right":keyboard.Key.right,
+        "caps_lock":keyboard.Key.caps_lock
+    }
 
-key_map = {
-    "shift":keyboard.Key.shift,
-    "ctrl":keyboard.Key.ctrl,
-    "up":keyboard.Key.up,
-    "down":keyboard.Key.down,
-    "left":keyboard.Key.left,
-    "right":keyboard.Key.right,
-    "caps_lock":keyboard.Key.caps_lock
-    } 
+    @classmethod
+    def parse_key(cls, key):
+        if key in cls.key_map:
+            return cls.key_map[key]
+        else:
+            return key
